chore: try recheck all comparissons

+++ Src/common/oscillation_generator.cpp
@@ -52,24 +52,26 @@ void SinSignalGenerator::setAmpl (uint16_t ampl) {
 void MultiSignalsSinGenerator::init() {
     signals_generator.resize(n_signals);
     uint16_t max_amplitude = 0;
-
-    for (int j = 0; j < n_signals; j++) {
+    
+    for (size_t j = 0; j < n_signals; j++) {
         std::uniform_int_distribution<uint16_t> freq_dist(min_freq, max_freq);
         std::uniform_int_distribution<uint16_t> amp_dist(1, 100);
         
         uint16_t freq_hz = freq_dist(rd);
         uint16_t amplitude = amp_dist(rd);
-
+        
         signals_generator[j] = SinSignalGenerator(sample_rate_hz, freq_hz, amplitude);
+        
         if (amplitude > max_amplitude) {
             max_amplitude = amplitude;
             dominant_sig.insert(dominant_sig.begin(), {amplitude, freq_hz});
         }
     }
-    std::stable_sort(dominant_sig.begin(), dominant_sig.end(), 
-    [](const std::pair<uint16_t, uint16_t>& a, const std::pair<uint16_t, uint16_t>& b) {
-        return std::less<uint16_t>{}(a.first, b.first);  
-    });
+    
+    std::stable_sort(dominant_sig.begin(), dominant_sig.end(),
+        [](const std::pair<uint16_t, uint16_t>& a, const std::pair<uint16_t, uint16_t>& b) {
+            return std::less<uint16_t>{}(a.first, b.first);
+        });
 }
